Move ssl_match_hostname to urllib3.utils.
