+++ noxfile.py
@@ -14,9 +14,6 @@ TYPED_FILES = {
     "src/urllib3/_collections.py",
     "src/urllib3/fields.py",
     "src/urllib3/filepost.py",
-    "src/urllib3/packages/__init__.py",
-    "src/urllib3/packages/ssl_match_hostname/__init__.py",
-    "src/urllib3/packages/ssl_match_hostname/_implementation.py",
     "src/urllib3/util/connection.py",
     "src/urllib3/util/proxy.py",
     "src/urllib3/util/queue.py",
