+++ setup.cfg
@@ -1,6 +1,6 @@
 [flake8]
 ignore = E501, E203, W503, W504
-exclude=./docs/conf.py,./src/urllib3/packages/*
+exclude=./docs/conf.py
 max-line-length=99
 
 [metadata]
