+++ setup.py
@@ -99,8 +99,6 @@ setup(
     license="MIT",
     packages=[
         "urllib3",
-        "urllib3.packages",
-        "urllib3.packages.ssl_match_hostname",
         "urllib3.contrib",
         "urllib3.contrib._securetransport",
         "urllib3.util",
