+++ src/urllib3/connectionpool.py
@@ -32,7 +32,6 @@ from .exceptions import (
     SSLError,
     TimeoutError,
 )
-from .packages.ssl_match_hostname import CertificateError
 from .request import RequestMethods
 from .response import BaseHTTPResponse, HTTPResponse
 from .util.connection import is_connection_dropped
@@ -40,6 +39,7 @@ from .util.proxy import connection_requires_http_tunnel
 from .util.request import set_file_position
 from .util.response import assert_header_parsing
 from .util.retry import Retry
+from .util.ssl_match_hostname import CertificateError
 from .util.timeout import Timeout
 from .util.url import Url, _encode_target
 from .util.url import _normalize_host as normalize_host
