--- src/urllib3/packages/ssl_match_hostname/__init__.py
@@ -1,3 +1,0 @@
-from ._implementation import CertificateError, match_hostname
-
-__all__ = ("CertificateError", "match_hostname")
