+++ src/urllib3/util/ssl_match_hostname.py
@@ -5,7 +5,13 @@
 
 import ipaddress
 import re
-import sys
+from typing import Dict, Tuple, Union
+
+# https://github.com/python/typeshed/blob/master/stdlib/2and3/ssl.pyi
+_PCTRTT = Tuple[Tuple[str, str], ...]
+_PCTRTTT = Tuple[_PCTRTT, ...]
+_PeerCertRetDictType = Dict[str, Union[str, _PCTRTTT, _PCTRTT]]
+_PeerCertRetType = Union[_PeerCertRetDictType, bytes, None]
 
 __version__ = "3.5.0.1"
 
@@ -80,7 +86,7 @@ def _ipaddress_match(ipname, host_ip):
     return ip == host_ip
 
 
-def match_hostname(cert, hostname):
+def match_hostname(cert: _PeerCertRetType, hostname: str) -> None:
     """Verify that *cert* (in decoded format as returned by
     SSLSocket.getpeercert()) matches the *hostname*.  RFC 2818 and RFC 6125
     rules are followed, but IP addresses are not accepted for *hostname*.
