+++ test/test_connection.py
@@ -4,10 +4,10 @@ from unittest import mock
 import pytest
 
 from urllib3.connection import RECENT_DATE, CertificateError, _match_hostname
-from urllib3.packages.ssl_match_hostname._implementation import (
+from urllib3.util.ssl_match_hostname import (
     CertificateError as ImplementationCertificateError,
 )
-from urllib3.packages.ssl_match_hostname._implementation import match_hostname
+from urllib3.util.ssl_match_hostname import match_hostname
 
 
 class TestConnection:
